requirements.txt
ragstack-ai
openai
flask
python-dotenv